@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -39,6 +39,7 @@ dev_dependencies:
 # The following section is specific to Flutter.
 flutter:
   assets:
+  - assets/lists.json
   - assets/items.json
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
