@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:main_app/controllers/ItemController.dart';
 import 'screens/Today Time Screen.dart';
+import 'controllers/List Controller.dart';
 
-void main() async{
+void main() async {
   runApp(MyApp());
 }
 
@@ -33,7 +35,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  MyHomePage({Key? key, this.title}) : super(key: key);
 
   // This widget is the home page of your application. It is stateful, meaning
   // that it has a State object (defined below) that contains fields that affect
@@ -44,16 +46,13 @@ class MyHomePage extends StatefulWidget {
   // used by the build method of the State. Fields in a Widget subclass are
   // always marked "final".
 
-  final String title;
+  final String? title;
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  
-
-
   @override
   Widget build(BuildContext context) {
     return TodayTimeScreen();
